requirements.txt
streamlit
python-dotenv
langchain
PyPDF2
chromadb
faiss-cpu
pdf2image
sqlite3